2025/02/19 commit (프로그래머스 <2021 카카오 인턴십> '표 편집')

+++ 250224/Programmers81303.java
@@ -46,7 +46,7 @@ public class Programmers81303 {
         	sb.append("O");
         }
         while(!stack.isEmpty()) {
-        	sb.insert(stack.pop().intValue(), "X");
+        	sb.insert(stack.pop(), "X");
         }
         return sb.toString();
     }
